Move new tests to a better place
